Move MockingExtensions to the SAGESharp.Testing namespace
