+++ SAGESharpTests/SAGESharpTests.csproj
@@ -78,7 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="MockingExtensions.cs" />
     <Compile Include="SLB\ASCIIExtensionsTests.cs" />
     <Compile Include="SLB\EqualityUtilsTests.cs" />
     <Compile Include="SLB\FooterEntryTests.cs" />
@@ -108,6 +107,7 @@
     <Compile Include="Testing\ParameterGroup.cs" />
     <Compile Include="Testing\AbstractEqualityByRefTests.cs" />
     <Compile Include="Testing\AbstractEqualityByValTests.cs" />
+    <Compile Include="Testing\MockingExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
