+++ SAGESharpTests/SLB/Level/Conversation/CharacterBinaryReaderTests.cs
@@ -2,6 +2,7 @@
 using NUnit.Framework;
 using SAGESharp.SLB;
 using SAGESharp.SLB.Level.Conversation;
+using SAGESharp.Testing;
 using System;
 using System.IO;
